WL#5593: Command-line utilities for MySQL: search database objects
Adding -e, --pattern option that can be used to supply a pattern.

Added check that a pattern supplied as first argument does not look like a connection specification. Patterns supplied with --pattern are not checked in this way.

Adding checks that no servers are supplied when using --sql option (they are ignored anyway).

Adding check that servers are supplied when --sql option is *not* used (the command will never match anything in that case).