WL#5615: Command-line utilities for MySQL: Documentation System
Adding the beginning of documentation using Sphinx. It contain
two man-pages (for mysqlprocgrep and mysqlmetagrep), documentation
for the 'grep' and 'proc' modules, as well as the beginning of an
documentation.

The patch also adds support to build and install manuals from
setup.py by using the commands 'build_html' and 'build_man'.
The 'install' command will install the manuals, but there is no
standard place to put the HTML documentation.