WL#5762 : mysqldbcheck
This patch introduces a sub component of the mysqldbcheck utility: the
mysqldiff utility. The mysqldiff utility allows users to check the
definitions among two databases for differences. It produces a diff-like
output in one of three formats.

Also included are documentation, new common module for use in 
mysqldbcheck, and tests for the utility.