WL#5753 : mysqlserverinfo
This patch adds a new utility that displays the common information about
a server. This is information needed for diagnosis or for configuration.
The utility has the option to show a list of running servers and the
default settings for mysqld from the configuration file. The user may
specify multiple servers to see information for a range of servers.

Also included are minor refactoring for the show servers code (moved 
from mut.py to server.py), and minor typographical corrects to 
mysqldiskusage.