BUG#12631924 : Parsing non-SQL files can result in wrong engine
This patch corrects a defect where mysqldbimport reading non-SQL files
(for example, CSV) that contain tables that have only one column can
result in the table having the wrong storage engine. The patch changes
the code to correctly save the last engine specified during the reading
of the import file. A test is added for regression.