BUG#12757358 : mysqldbcopy - duplicate % symbols inserted for SQL statements
This patch removes code that duplicates % signs in SQL CREATE
statements. The code was originally placed to avoid a problem with
string replacement but the code has evolved since then and this is
no longer needed. The code removed is responsible for the conditions
found in the bug report where % signs are duplicated.

A test has been added to verify this condition has been removed.