BUG#12885004 : mysqldbcopy, mysqldbexport - add --all option
This patch adds a new option, --all, to mysqldbcopy and mysqldbexport.
When the --all option is used, the utilities copy or export all of the
databases except for the system databases. This option is not valid
with a list of databases but can be used with the exclusion option.

The patch also corrects a minor omission in related tests concerning
proper shutdown of servers.

New test cases were added to ensure the --all operation works for
both utilities. The new error messages are also tested.