WL#5958 : show replication slaves
This patch includes the new utility mysqlrplshow which allows users to
show the slaves attached to a master. It will also show a replication
topology if the --recurse option is provided. The output is a graph of
the topology and an optional formatted list in either a GRID, CSV, TAB,
or VERTICAL format.