BUG#13082771 : mysqlserverclone - display command to start new server
This patch adds a new feature to mysqlserverclone to save the startup
command for the new server into a file with the --write-command option.

The patch also enhances the verbosity control of the utility thusly:

  * -v = normal verbosity
  * -vv = -v plus print startup command
  * -vvv = -vv plus debugging information

Tests also added in the same format as existing tests to include the
new feature.