WL#5988 :  MySQL Utilities: Add table locking to mysqldbexport and mysqldbcopy
This patch adds locking capabilities to mysqldbcopy and mysqldbexport. Users
can choose a consistent snapshot (default), no locking, or lock tables without
a transaction or consistent read.

Locking is made possible by a new Lock class added to the library as well
as an associated test to test the Lock class.

Tests updated to include test cases for locking options.

Also added are regexp search patterns for the --exclude option.