WL#5987 : add replication options to mysqldbcopy, mysqldbexport
This patch adds new options to the mysqldbexport utility to permit
exporting replication commands either to connect to the export server
as a master or to connect to the export servers master or both.

The patch also includes modifications to mysqldbimport to read and process
the replication commands.

The patch also includes new options for mysqldbcopy to allow stopping the
destination before the copy and connecting the destination to either the
source server as a master or the source servers master.

Additional replication tests are included to test the new features.