BUG#13721467 : Utilities throwing error on exit
When executing the MySQL Utilities on Windows with the latest
release of Workbench, some utilities throw an error on exit.
This is caused by not having an exit() call at the end of
the script to return an error code to the process that started
the script. This patch ensures all scripts have the exit()
call at the end to terminate as expected.