WL#6143 : Failover Utility for Global Transaction ID
This patch adds the mysqlrpladmin and mysqlfailover utilities in support of
the global transaction identifiers release of the server. These utilities
provide the user the ability to do switchover and failover on-the-fly
(mysqlrpladmin) or to monitor a replication topology and perform automatic
failover (mysqlfailover).

The patch includes changes to support running tests for the new utilities
as well as improvements in the tests for running on the 5.6.X servers
with and without global transaction identifiers enabled.