WL#6143 : Failover Utility for Global Transaction ID
This patch corrects the defects in the failover tests. The tests now
run on multiple platforms.

It also adds another extension point to the mysqlfailover utility to
permit identification of when a failover event has completed after the
logs have been updated, slave discovery has occurred (if selected),
and the health report has been regenerated.

Also adds terminal setting protection to MUT in case the mysqlfailover
console aborts too early. This restores the settings to the state
they were at the start of the test run.