BUG#13986713 : Wrong slave health output when slave uses loopback addr
When the slave connects to a master running on a localhost using the
loopback address, the health report can show the incorrect status of
the slave not connecting to the master. This is caused by the
comparison of 'localhost' to '127.0.0.1' which results in a 
mismatch when it should be equivalent.

This patch corrects the comparisons for slave connection, replication
user, and showing slave status. Tests included for regression testing.