BUG#14021531 : Replication user check fails for global host
If the replication user is created with rpl@'%', the mysqlrplcheck
test for the replication user will fail. This is because the code
is looking for the user 'rpl'@'somehost' and is ignoring the
'%' wildcard. 

This patch changes the code to include the '%' for host for the
replication user when checking to see if the replication user
exists and has the correct privileges.