BUG#14031894 : mysqlfailover fails to cleanup on error
The failover console can fail to deregister its instance on the
master should an error occur during processing. This patch adds
a try block around the console to ensure all exceptions are
caught and the console can deregister.

The patch also adds more clarity to the statements regarding
when a second instance of a console is started. It now tells
the user to restart the console with --force if starting a
new console after an event where the console was unable to
deregister (server went down, console crashed or was killed,
etc.).

Additional test refactoring included with an addition test
case to test killing of the console and restart.