BUG#14080657 : Slave election can fail if FILE vs TABLE used
The current code for the failover utilities includes a slave
election feature that has a defect. It can fail if the user
uses FILE output for master.info then later changes to TABLE
output. 

There was also a logic error that prevented the code from
working correctly unless verbosity was turned on. In some
cases the code can throw an exception and stop. This 
defect is in the code to check for the replication user on
the candidate slave.

This patch refactors the code for checking the replication
user and adds more information to error messages should
a candidate be eliminated.

The patch also updates the test files to include a test
case for the defect and related changes to support the
corrected code.