BUG#14021576 : exception thrown for GTID servers without password
In some cases where there are no passwords used, the check replication
user portion of the code can fail with an exception thrown. 

This patch corrects the defect along with correcting the associated
tests to ensure the correct replication user check is performed in
both cases where a password is and is not used.