BUG#14021551 : mysqlrplcheck not reading slave_master_info table
The replication code for utilities was not changed to read the
master information from a table per the --master-info-repository
startup option.

This patch adds this functionality to the replication modules
to permit the code to read the master information either from
a file or a table depending on the availability and setting
of --master-info-repository. Where =TABLE means to read from
the table, and either missing the variable master_info_repository
or a setting of =FILE to read from a file.

A new GTID-enabled test was created as well as existing GTID tests 
augmented to use (2) servers with --master-info-repository=TABLE
set.