BUG#14147098 : mysqlfailover fails if --force used on initial load
The --force option removes the registration table on the master and is used
in case the deregistration fails. However, if you start the console with
--force against a master that has never had a console registered, the
code will issue a query error. This was caused by --force attempting to
drop the table.

This patch adds the IF EXISTS clause to drop the registration table if and
only if it exists.