BUG#14159369 : fail|switchover can fail if --master-info-repository=file
The mysqlrpladmin and mysqlfailover utilities can fail to switcover
or failover if the slaves are using a file to store the master
information and the utility is run on a host that is not the slave
to which the switchover or failover is to occur.

This patch adds a new mechanism and requirement as follows.

- if the slaves do not have --master-info-repository=FILE, the
  use must provide the --rpl-user=<user>[:<password>] option
- if the --rpl-user option is specified, this user and password
  is used to setup replication among the new master and the
  remaining slaves

Additional test cases added to show the error condition described
above.