BUG#14158085 : connection fails to connect for aliases
When the failover utilities connection parameters are given using
an alias such as IP for hostname, loopback for localhost, etc. 
depending on how replication was started, the utilities may fail
to connect to or report that slaves are not connected.

The code for detecting if a slave is connected to a master must
be updated to include checking of aliases including, but not
limited to, IP, hostname, and alternative local host mappings.

Note: if IP is used and user does not have reverse DNS capability,
the connection test may still fail.