BUG#14238782 : serverclone does not ask before deleting data directory
This feature request removes the default behavior of automatically
deleting the new datadir specified by the --new-data option. 

This behavior was changed so that the  utility will only delete the
folder if it exists, is not empty, and the user specifies the
--delete-data option. 

If the folder exists and is not empty and --delete-data omitted,
the utility with present an error.

v2