Minor test updates - release-1.0.6
This patch includes minor test changes needed to comply with the
latest code in mysql-trunk. Among the minor modifications are
changes to how MUT and serverclone terminate servers. The new
server does not shutdown the same way as older versions and 
may require an additional step to send SIGTERM for the bootstrap
step. Also, a change to proc.py was made to ensure the output
is deterministic by sorting the output by operation.

Other minor changes are to tests for deterministic behavior
across platforms.