BUG#14298750 : slave_master_info not updated
This patch is a workaround for the problem where the server does
not write changes to the slave_master_info table when
--master-info-repository=TABLE is used.

The original workaround - to use --sync-master-info=1 and the
fact that the Connector/Python uses auto_commit=OFF by default
causes a locking issue when the CHANGE MASTER + START SLAVE is
issued from a single Connector/Python connection.

The solution therefore, indeed the workaround until the bug in
the server is fixed so that the slave_master_info table is
updated when CHANGE MASTER is issued, is to COMMIT the CHANGE
MASTER step. This patch adds a commit step after the CHANGE
MASTER command for setting up replication.