BUG#14378252 : replication user check fails with wildcard host
The internal check for the replication user and her privileges
can fail if the replication user is defnined with wildcards
for the host.

This patch corrects that limitation by permitting the search
to include wildcard matches in the same manner as the mysql
client resolves host wildcards.

A test case was added for regression.