BUG#14383189 : require --master-info-repo=TABLE for failover
This patch implements the feature request to require all slaves
to have --master-info-repo=TABLE for failover. This eliminates
a number of complex problems having to do with reading the
master information from a remote slave. Since failover requires
version 5.6.5 and later and crashproof slaves exist in that
version, the requirement does not limit functionality.