BUG#14501757 : mysqlrplshow -r includes spurious master
This patch corrects a defect in the discover slaves operation
in show replication topology. It now checks the slaves listed
in SHOW SLAVE HOSTS to ensure they are connected to the 
master.  

This introduces a new option for mysqlrplshow to discover
the slaves. The --discover-slaves-login option is required 
and requires a user:password pair in the same manner as
mysqlfailover and mysqlrpladmin.

It also includes clarifications in the documentation for
mysqlrpladmin and mysqlfailover regarding specifying slaves
with the --slaves option.

Test cases updated for regression.