BUG#14526940 : master listed on slaves list
This patch extends the check for checking to see if the master
is listed on the --slave, --slaves, or --candidates lists by
adding a check for hostname alias and port. This ensures if
the user specifies the master by IP the code will catch the
master listed as a slave by hostname.

Test cases added for regression.