WL#5752 : Utilities Console
This patch adds verbaige to the documentation to clarify how the
--execute option and user-defined variables work.

It also makes it possible for users to search for a utility by
omitting the 'mysql' prefix.

Lastly, it fixes an issue encountered when commands for the console
to execute are not placed inside quotes for the --execute option.