BUG#14711624 : mysqldbexport can fail if column is none
This patch adds the necessary check for None values before replacing single
quotes ('). In addition, single quotes are effectively replaced by the correct
escape sequence (\') instead of double quotes ('').

An unnecessary import was removed. In addition, the export_errors test case
was changed to support MySQL 5.5, and a rule was added to all mysqldbexport
test cases to require 5.1.0 servers or higher.