BUG#14711565: Exception thrown at Server module on unsupported server version.
An unhandled exception is thrown during the server version check if the server
instance is not created. This results in attempting to access a member of the
unrealized class instance. The Exception occurs only if the server is prior to
the version supported.

This patch includes changes to the message displayed to avoid the unhandled
exception. Three test cases was added to ensure there is no regression.