WL#6262 : Audit log parser
This patch fix some issues found during QA tests. In particular, it corrects
the rotate command of mysqlauditadmin, adds missing error checking for options
of mysqlauditadmin, and checks Python version requirements at the beginning of
the execution of both utilities (mysqlauditadmin and mysqlauditgrep). The
documentation was also updated in accordance to the referred changes.