BUG#13559593: mysqldiskusage crashes if --empty is specified and no empty dbs
This patch includes a fix for the issue of specifying the --empty option
and no empty (orphaned) databases exist. This was fixed by testing
the type of the data retrieved from the list of folders in the
datadir. A test case was added to show this has been fixed.

It also fixes a regression introduced with the SR#3-628710921 patch.
The change to detecting the version number failed to identify a
revision of 2 or more digits.

Also fixed is an issue with calculating disk size when verbosity is
turned on. The code was producing multiple rows. Now it prints one
row per database when -vvv is used.

v2