BUG#14624358 : mysqlfailover should report gtid information for master
This patch adds the GTID_executed_set from SHOW MASTER STATUS to the main
screen of the failover console. It also adds a new screen to the GTID
feature whereby the first screen is the list of the masters GTID executed
set split by GTID.

Note: if there are multiple GTIDs listed in the SHOW MASTER STATUS view,
the utility prints the first GTID and '[...]' to indicate the user
should check the GTID screens for the complete listing.

v2