WL#6261: Tarball and RPM distribution of MySQL Utilities
This patch adds or improves support for distributing MySQL Utilities using
RPM or archives like TAR or ZIP.

For the RPM distribution, we use a spec-file which an be found in the
support/RPM and replaced the bdist_rpm Distutils command
An RPM can be created on any system which has the rpmbuild tool and
also has Sphinx (for documentation) installed.

TAR and ZIP archives can be created using the sdist command. Formats 
can be specified using the option "--formats=gztar,zip".

Note that this patch changes the way the version is defined. We are now
using a more Pythonic style, for example (1, 2, 0, 'a', 0). The Workbench
version is mentioned in the description section of the RPM and the
PKG-INFO file.