BUG#14021584 : mysqlfailover fails to promote a slave when master shut down
BUG#14021584 was fixed by the patch for BUG#14802229. I.e., the "is_alive"
function in mysql/utilities/common/server.py was forgetting to set the result
to False when "is_connected" fails. This patch also fixes an issue found when
a down server (i.e. not running) is specified in the list given by the
--slaves option, ignoring not running slaves. Other corrections where made to
catch the correct exception when connecting to a server.

In addition, this patch updates the mysqlfailover documentation: correct the
supported MySQL server version and slaves options requirements.