BUG#15867353 : Add GTID handling to mysqldbcopy, mysqldbexport
This patch adds capability to include GTID handling for copy and export.
It generates the necessary commands to ensure the destination server has
the correct GTIDs once the copy or export/import is complete.

The statements include the following.

  # Turn off session binlog
  SET @MYSQLDUMP_TEMP_LOG_BIN = @@SESSION.SQL_LOG_BIN;
  SET @@SESSION.SQL_LOG_BIN= 0;
  # GTID state at the beginning of the backup
  SET @@GLOBAL.GTID_PURGED=<GTID_LIST;
  # EXPORT/COPY STATEMENTS GO HERE
  # Turn on session binlog
  SET @@SESSION.SQL_LOG_BIN = @MYSQLDUMP_TEMP_LOG_BIN;


The patch also include minor refinements found during testing of the
changes. These include:

  - Correcting the "top" message for mysqlindexcheck
  - Removal of unused methods in replication.py
  - Corrections to the import_engines test to limit execution
    to servers without sql_mode = NO_ENGINE_SUBSTITUTION
  - Removed non-cloning test cases from clone_db tests
  - Minor refactor to MUT to allow multiple skip- and do-test wildcards
    
v3