BUG#16010766 : gtid enabled utilities need to check version of the server
This patch adds a version check of the GTID feature to include checking
if GTID_MODE=ON, the server version is 5.6.9 or higher, and that the
gtid_executed system variable is present.

This is needed because interim builds of the server from 5.6.5 and 5.6.8
have an older feature set of GTID variables and are no longer compatible
with the current release of MySQL Utilities.

A test was added to exercise the new method added to perform the checks.

v2