BUG#16023781 : switchover can fail to complete if there are errors in slaves
The options --discover-slaves-login or --slaves must be used for the
mysqlrpladmin utility to be able to obtain the list of slaves. The utility now
check if either of the options is specified, if not an error is displayed and
the mysqlrpladmin terminates its execution.

The documentation was updated, adding the missing option requirement for the
switchover command. An unused import was also removed from the switchover
method.

Note: If there are errors in the replication stream (i.e. slaves), switchover
will continue if the --force option is used.