BUG#16020953 : --timeout option in mysqlfailover throws error
This patch fix an error that was thrown when the timeout option was used in
mysqlfailover. In particular, it checks if the specified value for the
--timeout option is valid and converts it to an integer.

In addition, the default timeout used in the function
WAIT_UNTIL_SQL_THREAD_AFTER_GTIDS (for the new master to catch with the
remaining slaves) was changed to a "high" value, in order to avoid the
function to terminate prematurely (without processing all transactions).

A test was added for regression and some corrections/improvements were made to
the failover tests.