BUG#13931340 : mysqluserclone should dump all users
This patch adds a new feature to mysqluserclone that permits administrators
to list all users with the --list option. If the user also specifies the
--dump option, the utility will also show the grants for all users.

The patch also includes a new feature for MUT that permits developers to
specify a test to stop executing enabling the execution of tests from
a given test to another test in sequence.

v3