BUG#16290692 : read port and socket information from login-path group
This patch enables the authentication mechanism using login-path to read the
port and socket from the .mylogin.cnf file when available and not specified
by the connection string. The values of the port or socket specified in the
connection string (<login-path>[:<port>][:<socket>]) will always overwrite 
the value read from the .mylogin.cnf file.

Note: The support for port and socket was only introduced as of MySQL 5.6.11
to the mysql_config_editor tool, therefore no MUT test was added for those
specific options as their availability depends from the version of the tool.