BUG#16164767 : mysqlreplicate fails to start slave
This patch improves error checking for mysqlreplicate. Now, the utility also
reports the SQL thread errors as well as if the IO and SQL threads are
running. In addition, if mysqlreplicate detects that the SQL thread is not
running while checking the slave status then it attempts to restart it.

Note: the issue is no longer reproducible with MySQL server 5.6.10 (GA).