BUG#16283330 : failover algorithm does not include wait for retrieved gtid set
This patch adds a new step to the failover method to ensure any transactions
read by a slave but not executed are captured by the new master.

The patch also adds a step to check each slave to see if it is necessary to
make the new master a slave. This avoids a potentially lengthy negotiation
when there are no events on the slave that are not on the new master.

Test included to test for each condition above.

Note: Due to new editor settings, trailing blanks are now stripped from
      those files changed by this patch. This conforms to the PEP standard.

v3