BUG#16238348: Python version not detected
This patch adds a check for the Python version to each utility and MUT.
It defines a minimum of version 2.6.0 and maximum version 3.0.0 (exclusive)
by default but the range can be requested for each check.

It will not raise Python exceptions, but an error message saying that the
Python version detected is not supported (only versions 2.5.0 and higher).