WL#5613 : .frm reader
This patch adds A new error handling exception and now permits the
use of the optional database name in paths. Also include are changes 
to the --help option to clarify the use of paths. The errors added are:

The new error is: --port must be different from the --server option.

The utility now accepts /home/me/data1/db1:t1.frm as a valid .frm
file reference. It will use db1 as the database name instead of 
data1 (or throwing an error).

v2