BUG#16243584: mysqlrpladmin stop, start, reset must require --slaves option
This patch adds a check for missing --slaves option for the start, stop,
and reset commands. It also changes the messages presented to the user
during the stop, start and reset commands. It also checks the slave
connection for the stop command and whether the slave is connected to
the master for the reset command.