BUG#16243584: mysqlrpladmin stop, start, reset must require --slaves option
This patch removes the restriction of not running the start, stop, or reset
command when there are warnings. The previous patch omitted running the
commands when there are warnings. This solution returns the code to the
original action where the command is run regardless if there is a warning.

The patch also fixes several test failures related to the original and
this corrected patch.

v1