BUG#16445510 : .frm reader does not execute as root
The utility currently prevents the user from running under a root
user account. This patch removes the restriction and adds a new
option named --user that allows the user to specify a user account
to use to execute the server. This allows users to run the utility
with root (e.g. su or sudo) and restrict execution of the spawned
server to a user account in accordance with best practices.

It is an error to attempt su or sudo without the --user option.

A test case was added to test the option.

v2