BUG#16202902 : mysqlrpladmin --master option is confusing
This patch corrects the option requirements for the failover command of
mysqlrpladmin. In particular, a warning is issued if the --master option is
used and its value is ignored. As the master information is no longer
available, an error is issued if the --discover-slaves-login option is used.
Therefore, the --slaves options is always required for the failover command.
Tests were added to MUT to validate this new requirements.

In addition, some extras fixes were made for MUT tests to run correctly and
regarding the options requirements for mysqlrpladmin, namely:
  - Fix the check of the required options for the commands 'elect', 'health' 
    and 'gtid'; 
  - Remove redundant options check; 
  - Standardize the error and warning messages format for all options check; 
  - Requirements for MUT tests rpl_admin_warnings and rpl_admin_errors have
    been changed;
  - Clean-up and refactor some MUT test modules (e.g. rpl_admin_warnings and 
    rpl_admin_errant_transactions);