BUG#16454212 : serverclone can change password of root user
On Windows platforms, it is possible when using mysqlfrm in default
mode with a port that is already in use by an existing MySQL server
that the root password will be changed.

This defect is located in the mysqlserverclone utility.

This patch adds a test case to show that the problem was fixed
in the mysqlserverclone utility.

v4 - 2 of 2