BUG#16454212 : serverclone can change password of root user
On Windows platforms, it is possible when using serverclone and
cloning a server to a port that is already in use by an existing
MySQL server that the root password will be changed.

In this case, the clone server bootstrap and start steps will
fail but no errors are returned thereby not halting the step to
change the root password.

This defect affects any utility that uses the serverclone
command module.

This patch corrects the defect by first detecting if the port
specified is free. If it is not, the clone server command
module fails before the bootstrap.

A test case was added to show the new check succeeds.

v4 - 1 of 2