BUG#13650863 : mysqldiff, mysqldbcompare not work on table names with '-'
This issue was fixed by the patch for BUG#14348501 (i.e. database objects
identifiers are now quoted with backticks).

Nevertheless, this patch fixes an issue with mysqldiff and mysqldbcompare when
databases with a '.' (dot) in the name are compared, for example a database
named `x.x`. MUT tests were changed to consider this situation, namely: the
data used to compare databases with weird names was modified to include a '.'
(dot) in the database name, and existing tests were corrected accordingly.

In addition, other MUT tests were also changed to run correctly independently
from the default database engine of the used server.