BUG#16210447 : failover does not work on unstable topology
This patch adds a check of the status of all slaves (before starting
failover), using SHOW SLAVE STATUS, reporting any error found and warning the
user that failover might result in an unstable replication topology. The
following default behaviors are provided:
 - For mysqlfailover, by default the process will not stop if errors are
   found, but if the --pedantic option is used then failover will stop with an
   error.
 - For mysqlrpladmin, by default the process will stop if errors are found,
   but if the --force option is used then failover will not stop and only
   issue a warning.
Note: IO errors on slaves are expected (prior to failover) since the master is
down, therefore only SQL errors will be checked.

A new MUT test case was created to test this new feature and existing tests
were corrected accordingly.

In addition, the find_errant_transactions method was improved to work when
some slaves are not alive, skipping them.