BUG#16489494 : mysqlrpladmin failover crashes when gtid is off
This patch solves the crash issue of the mysqlrpladmin utility when a slave
with GTID_MODE=OFF is specified for the failover command. Now, the utility
outputs an error message (instead of a warning) if some slave is found without
the GTID_MODE enabled and exits (cleanly). In particular, all the slaves with
the GTID_MODE=OFF are identified by the output error, and if some slave does
not support GTID it is marked as GTID_MODE=NO.

A MUT test was also added for regression.