BUG#16520505 : mysqlrpladmin switchover takes too much time to check slaves
This patch changes the input value used for retry when checking the slaves for
errors after the execution of the switchover (and failover) command of
mysqlrpladmin. Now, the value from the --ping option (by default 3) is used
instead of the --timeout value (by default 300).

Note: No MUT test was added because the condition to retry to check the slaves
status are supposed to occur only on unpredictable error circumstances, and
currently it is hard to design a reliable test that will deterministically
test this situation.