BUG#16268060 : mysqlrpladmin fail to start slaves with --discover-slaves-login
This problem cannot be reproduced because the use of the --discover-slave-
login option is no longer allowed for the start, stop and reset command,
instead the --slaves option is required. This was changed by the patch for
BUG#16243584. Nevertheless, this patch add some improvements to the
mysqlrpladmin utility for the start, stop and reset commands, namely:  

  - The --master option is no longer mandatory for the start, stop and reset
    command. When, it is not used the utility will not check if the specified
    slaves are configured for the specified master;
  - The options --discover-slaves-login and --slaves cannot be used 
    simultaneously for any command, otherwise an error is displayed and the 
    utility exits;
  - The output of slaves discovery (when --discover-slaves-login is used) was
    improved, and skipped slaves (with IO thread not running) are clearly 
    identified;
  - The verbosity option was added to mysqlrplshow. When the --verbose option
    is used the utility shows if the IO thread of the slaves is running. 

MUT tests were also added for regression.