BUG#13383767 : Utilities do not show version and copyright info
This patch adds a customization to the --help option that prints the
version and copyright banner for all utilities.

This required a small change to the utilities.py module for reading
help from the --help output for each utility.

The tests are masked for the version information since the version
will change for each release. A single test case was unaltered to
provide a regression test case.

v2