BUG#16210447 : failover does not work on unstable topology
This patch fix a regression issue that leads to unexpected errors when a
server specified in the slaves list is not acting as a slave in the topology.

This patch adds a check of the slaves status (before starting failover)
and also verifies if the slaves are not acting as such and issue an
appropriate warning or error message.