BUG#16204732: Replication Utilities are unable to use loopback addresses
This Bug fixes Replication Utilities to be able to handle the IPv4 and 
IPv6 lookback addresses (127.0.0.1 and [::1] respectively) instead of 
localhost as generic hostname.

Includes a change in MUT to clone servers using lookback addresses as 
host instead of localhost and regression tests for IPv4 and IPv6 lookback 
addresses.

v2: fixes a bug introduced on previews added test