BUG#16210222: switchover fails due to missing or wrong password.
This patch fixes the failing cases, when master_repository_info is set to 
FILE and replication user's password is required, but the user specify an 
incorrect password or none at all. Resulting on servers not connecting to 
new server.

The patch corrects the behavior as follow:
  - when password is not specified and --force is not used, an error will 
    be show to the user, however if --force the password will be cleared 
    for the specified user.
  - when a wrong password is specified and --force  not is used, an error 
    will be shown to the user, however if --force the password will be 
    overwrote for the specified user.

A regression test was created to test these cases.

v2