BUG#16410648: mysqldiff is unreliable
This patch adds a fix to mysqldiff that it finds differences even though no
difference exists. This happens on tables when the order of the columns or
indexes are different.

Now, the order of columns is not taken into account in the comparison if the
option --changes-for is omitted. In this case, if the --quiet option is not set
a WARNING will be displayed to inform that the tables structure is the same,
but the order of some definitions is different.

If the option --change-for has a value and column order is different, the
tables are different and it will show the ALTER TABLE statement
if --difftype=sql.

The order of indexes in the tables definition is ignored, because it has no
influence on the tables.