BUG#16584598 : mysql switchover has no proper error with wrong slave name
This patch adds a check to verify if the specified slaves (option --slaves) are
configured for the specified master (option --master) when the switchover
command is executed. By default, switchover will output the error and stop,
unless the --force option is used to ignore the error and force the operation
to proceed (i.e. not stopping).

In addition, an issue was also corrected when offline servers are included in
the --slaves list.

MUT tests were added for regression, and some current tests were fixed.
Namely, servers were killed in tests cases that create unstable topologies to
avoid unexpected issues in following the tests.