BUG#16565502: mysqlrpladmin switchover new master is actual master
This bug fixes the case where mysqlrpladmin fails if the new master 
specified is the original master. The behavior has been changed so that 
if this occurs, an error is issued. 

The error messages for the case above, and when the master and one of 
the candidates or slaves are the same, has been improved.

The server version requirements for rpl_admin test have been change 
from version < 5.6.6 to version > 5.5.30 and GTID not enabled.

v3b