BUG#16589086 : mysqlreplicate crashes when no options are used
This patch prevents mysqlreplicate and mysqlrplcheck from crashing when no
option is specified displaying an error message with the missing option. In
addition, a warning message is displayed when the --master option is not
specified to notify the user of the value used by default.

MUT tests were added for regression.

Note: function 'run_test_case' in the class replicate.test was renamed to
'run_rpl_test' to avoid being confused with the function from the superclass
System.test with the same name.